Rebalanced Health
I added and removed many features such as, title movement, dashing, and sprinting. These changes were all scrapped due to being too difficult to add at my current skill level and the time frame given. The only thing that was changed and stuck was the change to the spawning of the health power as now it only comes once a minute making it more of a saving last will rather than a constant rate.